+++ libs/cardano-ledger-core/testlib/Test/Cardano/Ledger/TreeDiff.hs
@@ -304,6 +304,12 @@ instance ToExpr (SnapShots era)
 
 instance ToExpr SnapShot
 
+instance ToExpr MarkSnapShot
+
+instance ToExpr SetSnapShot
+
+instance ToExpr GoSnapShot
+
 instance ToExpr StakePoolSnapShot
 
 deriving newtype instance ToExpr Stake